refactor(tests): update import paths for binary, file, and image field test files
### Summary

- Adjusted import statements in `binary-field.spec.ts`, `file-field.spec.ts`, and `image-field.spec.ts` to reference the correct relative path for the `file-test-data` helper.
- This change improves the organization of test files and ensures consistency in import structure across the test suite.

### Checklist
- [x] Updated import paths for clarity and maintainability

This refactor enhances the structure of the E2E tests without introducing any functional changes.
